# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CourseLevel } from "./courseLevel";

export interface Course {
  id: number;
  title: string;
  description: string;
  category: string;
  instructor: string;
  level: CourseLevel;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  durationHours: number | null;
  isPublished: boolean;
  createdAt: Date;
  updatedAt: Date;
}