# lib/api-zod/src/generated/types/courseStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CourseStats {
  courseId: number;
  title: string;
  category: string;
  enrollmentCount: number;
  completionCount: number;
  completionRate: number;
}