# lib/api-zod/src/generated/types/enrollment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EnrollmentStatus } from "./enrollmentStatus";

export interface Enrollment {
  id: number;
  courseId: number;
  studentName: string;
  studentEmail: string;
  status: EnrollmentStatus;
  enrolledAt: Date;
  /** @nullable */
  completedAt: Date | null;
  progressPercent: number;
  /** @nullable */
  courseName: string | null;
}