# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Lesson {
  id: number;
  courseId: number;
  title: string;
  content: string;
  /** @nullable */
  videoUrl: string | null;
  /** @nullable */
  durationMinutes: number | null;
  order: number;
  createdAt: Date;
  updatedAt: Date;
}